#12 added new slice counter
